+++ decompiler/ff7_field/ff7_field_disassembler.h
@@ -319,7 +319,7 @@ namespace FF7
         void DisassembleIndivdualScript(std::string entityName,
             size_t entityIndex,
             size_t scriptIndex,
-            int16 scriptEntryPoint,
+            size_t scriptEntryPoint,
             uint32 nextScriptEntryPoint,
             bool isStart,
             bool isEnd);
